# openapi.json
{
  "components": {
    "schemas": {
      "ActivationCreate": {
        "properties": {
          "config_cid": {
            "title": "Config Cid",
            "type": "string"
          },
          "expected_head": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "CID (hex) of the activation the client believes is the current head; null for the first activation. Stale heads are rejected (409).",
            "title": "Expected Head"
          }
        },
        "required": [
          "config_cid"
        ],
        "title": "ActivationCreate",
        "type": "object"
      },
      "ActivationRead": {
        "properties": {
          "agent_uuid": {
            "format": "uuid",
            "title": "Agent Uuid",
            "type": "string"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "config_cid": {
            "title": "Config Cid",
            "type": "string"
          },
          "prev_cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Cid"
          }
        },
        "required": [
          "cid",
          "agent_uuid",
          "config_cid",
          "prev_cid"
        ],
        "title": "ActivationRead",
        "type": "object"
      },
...
      "ConfigCreate": {
        "properties": {
          "llm_config": {
            "additionalProperties": true,
            "title": "Llm Config",
            "type": "object"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          },
          "tools": {
            "items": {},
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "system_prompt"
        ],
        "title": "ConfigCreate",
        "type": "object"
      },
      "ConfigRead": {
        "properties": {
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "llm_config": {
            "additionalProperties": true,
            "title": "Llm Config",
            "type": "object"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          },
          "tools": {
            "items": {},
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "cid",
          "system_prompt",
          "llm_config",
          "tools",
          "metadata"
        ],
        "title": "ConfigRead",
        "type": "object"
      },
      "DUTCreate": {
        "properties": {
          "agent_output": {
            "title": "Agent Output",
            "type": "string"
          },
          "agent_uuid": {
            "format": "uuid",
            "title": "Agent Uuid",
            "type": "string"
          },
          "artifact_cids": {
            "description": "hex CIDs of the byte artifacts touched",
            "items": {
              "type": "string"
            },
            "title": "Artifact Cids",
            "type": "array"
          },
          "business_object_keys": {
            "items": {
              "type": "string"
            },
            "title": "Business Object Keys",
            "type": "array"
          },
          "config_cid": {
            "description": "hex CID of the behavioral AgentConfig used",
            "title": "Config Cid",
            "type": "string"
          },
          "input_context": {
            "title": "Input Context",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "self-asserted event time; defaults to server now",
            "title": "Occurred At"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "span_id": {
            "description": "runtime span handle, assigned at span start",
            "title": "Span Id",
            "type": "string"
          }
        },
        "required": [
          "agent_uuid",
          "config_cid",
          "span_id",
          "sequence",
          "input_context",
          "agent_output"
        ],
        "title": "DUTCreate",
        "type": "object"
      },
      "DUTRead": {
        "properties": {
          "agent_cid": {
            "title": "Agent Cid",
            "type": "string"
          },
          "agent_output": {
            "title": "Agent Output",
            "type": "string"
          },
          "artifact_cids": {
            "items": {
              "type": "string"
            },
            "title": "Artifact Cids",
            "type": "array"
          },
          "business_object_keys": {
            "items": {
              "type": "string"
            },
            "title": "Business Object Keys",
            "type": "array"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "config_cid": {
            "title": "Config Cid",
            "type": "string"
          },
          "input_context": {
            "title": "Input Context",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "span_id": {
            "title": "Span Id",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "span_id",
          "business_object_keys",
          "sequence",
          "input_context",
          "agent_output",
          "occurred_at",
          "agent_cid",
          "config_cid",
          "artifact_cids"
        ],
        "title": "DUTRead",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "aits",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/agents/{agent_uuid}/activations/": {
      "get": {
        "operationId": "configs_activation_history",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivationRead"
                  },
                  "title": "Response Configs Activation History",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activation History",
        "tags": [
          "configs"
        ]
      },
      "post": {
        "operationId": "configs_activate",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate",
        "tags": [
          "configs"
        ]
      }
    },
    "/agents/{agent_uuid}/active-config/": {
      "get": {
        "operationId": "configs_active_config",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Active Config",
        "tags": [
          "configs"
        ]
      }
    },
    "/agents/{agent_uuid}/config/": {
      "put": {
        "description": "Make this config the agent's active one, in one call: record the\nbehavioral content (dedup) and activate it against the server-resolved\nchain head. Setting the already-active config is a no-op. Callers that\nneed the compare-and-set guard use ``POST .../activations/`` instead.",
        "operationId": "configs_set_config",
        "parameters": [
          {
            "in": "path",
            "name": "agent_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Config",
        "tags": [
          "configs"
        ]
      }
    },
...
    "/configs/": {
      "post": {
        "operationId": "configs_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "configs"
        ]
      }
    },
    "/configs/{cid_hex}/": {
      "get": {
        "operationId": "configs_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "configs"
        ]
      }
    },
    "/duts/": {
      "post": {
        "operationId": "duts_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DUTCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DUTRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "duts"
        ]
      }
    },
    "/duts/{cid_hex}/": {
      "get": {
        "operationId": "duts_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DUTRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "duts"
        ]
      }
    },
...
  }
}